Renamed module
The public functions are all part of the AsciiArt struct, so it makes more sense to name it `ascii_art` rather than `parse`
